completed ATM/confirm
added processing.py and methods for each transaction type
after POSTing a form that has a key 'TransactionType' in the POST dictionary with a valid value('Withdraw', 'Transfer', 'Enquiry'). The confirm view asks the appropriate method in processing to perform the transaction and return a message to display in the confirmation screen.